Improve camera rig
I was finding the linear movements of the camera positioning and camera look limiting, so I expanded the camera "rig" to include a base that can rotate around the center. This helps get more of the effect I was looking for.

An alternative might have been to just move the models around appropriately, but this would have cause a shift in the lighting which would not have been desirable.