Extract three content and refactor
This makes suer that we now set the position on each frame as opposed to only on scroll (though the values are only updated on scroll).

Since I can now use R3F hooks inside this new component as well, that hopefully opens the door for some more responsive code by being able to leverage viewport sizing and such.